requirements.txt
huggingface_hub==0.29.2
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.8.5 